+++ src/navigation/StackAfterLogin.js
@@ -6,7 +6,7 @@ import SignOut from '../screens/mypage/SignOut'
 import SignOutFinish from '../screens/mypage/SignOutFinish'
 import Search from '../screens/main/Search'
 import SearchResult from '../screens/main/SearchResult'
-import RecruitDetails from '../screens/main/recruit/RecruitDetalis'
+import RecruitDetails from '../screens/main/participate/RecruitDetalis'
 import Useredit from '../screens/mypage/Useredit'
 import Phoneedit from '../screens/mypage/Phoneedit'
 import Passwordedit from '../screens/mypage/Passwordedit'
